% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyvader.R
\name{get_vader_dictionaries}
\alias{get_vader_dictionaries}
\title{Get VADER dictionaries in a nested tibble.}
\usage{
get_vader_dictionaries()
}
\value{
A tibble containing all VADER dictionaries and canonical test
sentences.
}
\description{
This function returns a nested tibble containing all of the dictionaries used
to calculate VADER scores and canonical test sentences, so that users can
inspect them and understand what's happening under the hood. At present
end-users have read-only access to dictionaries; there is no way to add
custom dictionaries or make changes.
}
