% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyvader.R
\name{vader}
\alias{vader}
\title{Tidy VADER}
\usage{
vader(data, var)
}
\arguments{
\item{data}{A dataframe or tibble with a column of sentences to score.}

\item{var}{The name of the column containing the sentences.}
}
\value{
The original input with four additional columns for each sentence's VADER scores: compound, pos, neu, neg.
}
\description{
Tidy VADER
}
