% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyvader.R
\name{vader_chr}
\alias{vader_chr}
\title{VADER for a character vector}
\usage{
vader_chr(text)
}
\arguments{
\item{text}{A sentence to be scored, in a single string (i.e. a non-empty length-1 character vector).}
}
\value{
A one-row data.frame with four columns for the sentence's VADER scores: compound, pos, neu, and neg.
}
\description{
VADER for a character vector
}
